refactor: (#685) 모바일 기본 헤더를 레이아웃에 포함 및 적용
