+++ frontend/src/components/common/Layout/index.tsx
@@ -1,21 +1,28 @@
 import { PropsWithChildren, useEffect } from 'react';
+import { useNavigate } from 'react-router-dom';
 
 import ChannelTalk from '@components/ChannelTalk';
 import Dashboard from '@components/common/Dashboard';
 import WideHeader from '@components/common/WideHeader';
+
+import IconButton from '../IconButton';
+import NarrowTemplateHeader from '../NarrowTemplateHeader';
 
 import * as S from './style';
 
 interface LayoutProps extends PropsWithChildren {
   isSidebarVisible: boolean;
   isChannelTalkVisible?: boolean;
+  isMobileDefaultHeaderVisible?: boolean;
 }
 
 export default function Layout({
   children,
   isSidebarVisible,
+  isMobileDefaultHeaderVisible = true,
   isChannelTalkVisible = true,
 }: LayoutProps) {
+  const navigate = useNavigate();
   useEffect(() => {
     isChannelTalkVisible ? ChannelTalk.showChannelButton() : ChannelTalk.hideChannelButton();
   }, [isChannelTalkVisible]);
@@ -30,6 +37,18 @@ export default function Layout({
           <S.DashboardWrapper>
             <Dashboard />
           </S.DashboardWrapper>
+        )}
+        {isMobileDefaultHeaderVisible && (
+          <S.MobileHeaderWrapper>
+            <NarrowTemplateHeader>
+              <IconButton
+                category="back"
+                onClick={() => {
+                  navigate(-1);
+                }}
+              />
+            </NarrowTemplateHeader>
+          </S.MobileHeaderWrapper>
         )}
         <S.MainContainer $isSidebarVisible={isSidebarVisible}>
           <S.ChildrenWrapper $isSidebarVisible={isSidebarVisible}>{children}</S.ChildrenWrapper>
