+++ frontend/src/pages/VoteStatisticsPage/index.tsx
@@ -1,12 +1,10 @@
 import { Suspense } from 'react';
-import { useNavigate, useParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 
 import ErrorBoundary from '@pages/ErrorBoundary';
 
-import IconButton from '@components/common/IconButton';
 import Layout from '@components/common/Layout';
 import LoadingSpinner from '@components/common/LoadingSpinner';
-import NarrowTemplateHeader from '@components/common/NarrowTemplateHeader';
 
 import OptionWrapper from './OptionWrapper';
 import StatisticsWrapper from './StatisticsWrapper';
@@ -15,15 +13,9 @@ import * as S from './style';
 export default function VoteStatisticsPage() {
   const params = useParams() as { postId: string };
   const postId = Number(params.postId);
-  const navigate = useNavigate();
 
   return (
     <Layout isSidebarVisible={true}>
-      <S.HeaderWrapper>
-        <NarrowTemplateHeader>
-          <IconButton category="back" onClick={() => navigate(-1)} />
-        </NarrowTemplateHeader>
-      </S.HeaderWrapper>
       <S.Container>
         <S.PageHeader>투표 통계</S.PageHeader>
         <S.ContentContainer>
